Try to fix LocaleSupportDaemonIntegrationTest flakiness
The problem for flakiness is that daemon with "release" version is still
alive and blocks deleting of logs file after tests on Windows platform.

The idea here is that we make daemon isolated and `AbstractGradleExecuter` aware of
the modified version. And with that daemon is hopefully killed successfully in
`AbstractGradleExecutor#cleanupIsolatedDaemons`, that is called from the `cleanup`
method at the end of the test.