Add simpler API to capture external process output as build input
There are two new methods, `exec` and `javaexec` in the ProviderFactory
class. These are similar to the existing methods of ExecOperations and
reuse the same configuration DSL classes. The returned providers have
convenient methods to read the process' output during the configuration
time or to wire it as an input to the task. The outputs become parts of
the configuration cache fingerprint if used during the configuration
time, so the wrapped process is re-executed every time build is started.