Respect working dir set on the exec/javaexec output Provider
It wasn't captured in the ValueSource parameters so default working
directory was used instead.

I verified that all other properties on ExecSpec/JavaExecSpec boil down
to command line modifications so there is no pressing need to capture
them separately. We might reconsider this if eager command-line
computation becomes a problem.