Only check moved roots of watched snapshots on Linux
and not for all watched paths. This is to improve
performance when there are many files.

We don't need to check if the path is still watched for files
inside of watched snapshots, since we also watch the parent directory
of the watched path, so we'd already get a file system event.