Fix the test expectations after adding input stream tracking
The instrumentation that tracks build logic inputs doesn't care if the
input access happens in the ValueSource implementation. It is most
likely overzealous, because ValueSource is going to be re-evaluated at
the next build anyway and may cause false invalidations (e.g. if the
ValueSource's value isn't affected by this particular change to its
detected input).

This doesn't affect correctness though, and requires a non-trivial fix
so we alter the test instead.