Fix flakiness in destroyer command line ordering test
There were two issues here:
* In the multiple destroyers test, the mechanism we use for synching the tasks via `BlockingHttpServer` also
releases the project lock.  This allows the other task in the same project to sometimes jump in and run
concurrently while it is calling the http server to release the sync.  So although the task we expect to run
first does run first, it can sometimes complete _after_ the second task, throwing off the expectations.  The
expectations have been updated to account for this possibility.
* In the destroyer-in-a-different-project test, the destroyer wasn't actually destroying the same directory
as the producer, which allowed them to sometimes run in parallel, throwing off the expectations.  Now that
they destroy the same directory, the producer should always run after.