Add --show-version switch
Prior to this commit gradle could only print the version or perform another task.
This commit adds a new set of flags '-V', '--show-version' which print the version and continue on with the regular execution.
To achieve this, the interface ComposableAction is added which can be returned from CommandLineAction.createAction instead of a plain runnable,
to indicate that the action can be composed with others.

Signed-off-by: Leonard Brünings <leonard.bruenings@gradle.com>