Use computeIfAbsent/putIfAbsent instead of repeated Map operations
This means that hash() is only called once instead of 2-3 times in
the previous code. I've also used cheaper keys where possible, e.g.
using ImmutableList instead of Arrays.asList. In some cases the object
was cheap enough to be its own key, a.k.a. we're not caching the to
avoid its creation cost, but to guarantee uniqueness.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>