Remove complexity around looking at http.nonProxyHosts
getProxy(host) was only used by S3 repositories, but we incorrectly looked
for https.nonProxyHosts for HTTPS connections. This means that we would
skip using the proxy for hosts that matched https.nonProxyHosts.

Except, internally, AWS's SDK looks at http.nonProxyHosts, which is the
standard/normal Java system property.

So if you had a proxy configured, a S3 host you wanted to use over HTTPS
and you wanted to _not_ use the proxy, you might set:

https.nonProxyHosts=example.com

which would cause us to _not_ provide proxy info to the S3 library, but
internally, the AWS SDK would still see the proxy settings and end up using
the proxy. Our documentation does not say to use this property and this
property is not mentioned anywhere.

So if you noticed this and really didn't want to use the proxy, you would
need to set the correct property:

http.nonProxyHosts=example.com

And this would start skipping the proxy for this host.

It looks like all of the logic around passing the proxy to S3Client
might be unnecessary because it's reading the same system properties Gradle
is, but this change doesn't try to remove more than necessary to remove
the unnecessary and confusing behavior around https.nonProxyHosts