Align memory limit with other tests
It seems that the test itself is fine but it no longer fits into 128M
because other parts of Gradle got bigger. There is no indication of any
leaks in GC logs even when running the test with >100 iterations.

The original failure seems to be triggered by the lack of Metaspace
anyway.