Allow certain kinds of daemon connection failures on older versions of Gradle
On older versions of Gradle, we can sometimes encounter problems when first
connecting to a Gradle daemon. This causes the test to fail with a "tried to
connect to 100 daemons" error. This appears to be fixed in later versions of
Gradle.

It looks like we see this error for <5.0, but I'm unsure when it was actually
fixed.