Keep specific version when restoring realized/serialized module resolve metadata
Gradle 7.4 and before...

Without serialization, the artifact's file name is based on a non-SNAPSHOT version.
With serialization, the serialized artifact was recreated with a SNAPSHOT version, but
in MavenResolver, we would ignore the artifact and create a new artifact with the
right version through FixedComponentArtifacts.

Since then, other changes have made MavenResolver rely on MetadataSourcedComponentArtifacts
completely and this has caused failures in MavenSnapshotResolveIntegrationTest due to
this mismatch between the serialized/non-serialized artifact metadata.