Ignore "java.awt.headless" in the configuration cache
Some popular plugins modify this property at runtime. This breaks
configuration caching if some other plugin reads the modified value -
the fingerprint check at the next run expects the property to have
the modified value so it fails and invalidates the cache.

This change doesn't fix the issue completely but reverts to the old
behavior: use the default (unmodified) value of the property for the
cached run. It is incorrect but for this particular case seems to be
an acceptable compromise. The proper fix is being tracked at #18432.

Fixes #19710.