Allow configuring rich dependency constraints using Kotlin DSL
This switches the type used by the methods provided in the dependencies block to use Actions based on
Module Dependencies, rather than Dependency.  This lets Kotlin understand adding configuration like exclude
blocks to them, which previously failed to compile.

The downside is that if you try to configure a non-Module Dependency (files, for example, or any other SelfResolvingDependency)
in this block, we can't handle it.  So throw an exception explaining this.

This also adds a good amount of test converge for dependency scenarios related to test suites.