Add support for dependency constraint without version
Prior to this change, a published Gradle Module Metadata that contained
a constraint without a version definition would result in a null
VersionConstraint instead of an empty one.
This caused issue during metadata serialization and potentially
resolution.
The field is not nullable and an empty version constraint is now used as
the default value.

Fixes #20189