Make memory settings for to checkstyle worker configurable
With having ported the Checkstyle task to use the worker api we see OOM problems
with our elasticsearch build when using the latest gradle master (7.5 snapshot)
This PR makes the heap size for checkstyle tasks configurable to not be forced to
rely on jdk defaults.
The property namings are similar to how memory settings are named in the Test task.

Signed-off-by: Rene Groeschke <rene@elastic.co>