Do not capture whole test instance when provider is enough
The test listeners live for the entire test run duration. Capturing
a test instance in the listener prevents it from being GC'ed, and the
test instance can be quite large. Having hunderds of them can easily
cause OutOfMemoryError.

This commit reduces the retained size of the listener by only capturing
the test directory provider which is much smaller.