Extract a type to represent the path through which a node is reachable from the various "root" nodes such as the requested tasks or a finalizer.
The path to a node can affect how the node is scheduled relative to other nodes, for example a node that is reachable only via a finalizer cannot start until the finalized task(s) have completed and cannot be cancelled if the finalized tasks(s) have completed.

Using a type to represent this logic and state moves it into a single place and out of the main node scheduling logic.