Add default toolchains for MacOS.
Fixes #1096

Adds two new default toolschain targets for MacOS x86_64 and
arm64 which use Clang's `-arch` argument to target the correct
architecture.  On MacOS, this is used in preference to the
existing `Intel64Architecture` target which passes `-m64` to the
toolchain.

This is sufficient for `NativeExecutableSpec` to target
`osx_x86-64` and `osx_aarch64` on both Intel and M1 Macs.

I haven't (yet!) located any tests for this functionality, and
so this PR doesn't include any test changes.  Happy to add them
if you can point me in the right direction.

Signed-off-by: Pete Bentley <prb@google.com>
Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>