Add providers to look up env vars and system properties by prefix
There are many user requests to be able to look up a subset of the
environment variables or system properties in configuration-cache
compatible way. Filtering the variables in the build logic makes all
them, even the unrelated ones, inputs to the configuraton cache,
reducing cache hits. The configuration cache machinery is unable to
understand the meaning of the filtering code in the build logic and
therefore cannot reduce the number of inputs.

This commit adds a specific API to address the most common use case
of selecting a subset of variables with names sharing a common prefix.
When the provider returned by the new API is resolved at the
configuration time, all variables with the prefix become an input.
No other variables affect the cache fingerprint.