Workaround for srcDistribution causing up-to-date jar tasks
This moves the source distribution tasks after
the bin distribution tasks, so it doesn't run
in parallel with the source distribution tasks
which may cause incorrect builds.

So the general problem is again that if some task C reads the output directory
after the task action of a task P started, then we may read stale data from the
VFS when task P finishes, since we only invalidate the output of P at the start
of P, but not just before we snapshot them after P did execute.

So in our case, we have the task C: `:distributions-full:srcDistributionZip`,
https://ge.gradle.org/s/akx6ljbtairda/timeline?details=ox4zjugq5bm4q&end=1649037234419&start=1649037234409
And P is `:execution:jar`:
https://ge.gradle.org/s/akx6ljbtairda/timeline?details=xqkf55kdgcruq&end=1649037234419&start=1649037234409
The trick here is that srcDistributionZip does not declare all of the
directories as an input, but only some filtered part:
https://github
.com/gradle/gradle/blob/45e8762d131f22fdf2c11eb528a93e7d24afc51d/build-logic/packaging/src/main/kotlin/gradlebuild/packaging/GradleDistributionSpecs.kt#L91-L116
Normally, we’d detect that we didn’t include the subprojects/execution/build
directory and then see that the snapshot isn’t complete. Though if none of the
:execution subproject tasks did run, that means that Gradle may detect the
directory as complete - there isn’t a build directory so nothing has been
excluded. And if the read starts before the task in the execution project run
and finished after they ran than we have a problem: We have the wrong state in
the VFS and think it is up-to-date. We also ignore file system events for those
locations, since they are part of the outputs. All in all, Gradle would think
that the JAR task did output nothing.

Coming back to the build scan above:
`:distributions-full:srcDistributionZip` started at 7.683s and it took 0.494s to
finish input snapshotting - which puts the end of input snapshotting at 8.177s
`:execution:jar` started at 8.070s and took 0.036s which puts it end at 8.106s.
While 8.177s is after 8.106s, it may still be that the inputs have been stored
in time in the VFS before the jar task finished causing it to think that it
didn’t produce any outputs.