Offer a more conventional default package name in build-init questionnaire
When the user runs the init task, the default value for the package name is the name of the project.
The project name, however, has no formatting enforced and can start with an upper-case character.
Upper-case package names are unconventional in Java, and even cause problems with test filtering,
as reported at https://github.com/gradle/gradle/issues/20350.

This change improves the situation by offering the project name lower-cased as the default value for
the package name in the init task command input questionnaire.

Fixes #20350