Add tests and document a relative order of onAccess/onChange
The relative order of onAccess and onChange/onRemove is important to the
configuration cache because accesses to changed properties are no longer
tracked, therefore the information about initial access (because e.g. of
the return value of the putProperty) may be lost if the order is
incorrect.

However, the state of the Properties object for the duration of the
callback is not important and it is hard to impose a guarantee there, so
it is only documented to be undefined.