Do not run some UndeclaredBuildInputsIntegrationTest in no-daemon mode
The test has to change the Gradle command line to inject a system
property. In the no-daemon mode the change of the command line affects
"sun.java.command" system property that holds this command line. In
all other modes the changes doesn't affect the original command line.
By the nature of the tests all system properties are inputs to the
configuration cache and the no-daemon behavior unexpectedly invalidates
the cache.

This test doesn't test anything particularly specific to no-daemon or
daemon mode and can be safely skipped in the no-daemon configuration.