Make external processes at the configuration time fail build only if...
...the STABLE_CONFIGURATION_CACHE feature is enabled.

Some popular plugins start external processes at the configuration time
and haven't had time to adapt. Making the process start a hard failure
can hurt the user's ability to adopt the configuration cache. This
commit hides the error behind a feature flag, so the build and plugin
authors can enable stricter validation.

One downside of the feature flag-based approach is that it doesn't allow
to cover init scripts and settings plugins, as these are
evaluated/applied before the actual settings script is executed.