Always set TD-specific server to avoid creating conflicting clients
When PTS is enabled but TD is disabled, the TD Gradle plugin creates a
client service with the server configuration of that test task. If a
later test task enables TD with a different server, an exception is
thrown. This failure is now avoided by always configuring the
TD-specific server, even if TD is disabled.