Fix composite build sample documentation
This sample has gone out of date and no longer executes properly when running
the steps in the documentation as directed. In order to get the sample working
correctly, the following changes were made:
1) Update all local publishing to use maven instead of ivy
2) Update commons-lang dependency to 3.12.0 from 3.4. The directions tell the user
   to clone the commons-lang repo, but said to clone the  branch. The mainline
   code was updated to use non-ascii encoding for the java files, which exposed a bug
   in the way we convert maven POMs to gradle files when running . So,
   we update the readme to instruct the user to explicitly clone the 3.12.0 branch
   so future changes to commons-lang will not affect this sample.
3) Remove unused repository declarations
4) Update Maven2Gradle conversion code (and tests) to ensure the javadoc task is also
   informed of file encoding instead of only the JavaCompile tasks.
5) Update README to direct the user to run the  task instead of just .
6) Use a modern git repo URL for the commons-lang repo.

Additionally, in order to have the sample project not escape the sample directory
and modify external files, the location of the  was moved one directory
down.

Tested by running  with the Maven2Gradle changes and running through the
updated procedure with the sample zips output after these changes.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>