w dependencies with explicit artifacts to skip metadata fetching
Allow dependencies declared with an explicit artifact (classifier or
extension) to skip metadata fetching (POM/IvyDescriptor). This was
implemented by updating the `Default[Ivy,Maven]ArtifactRepository`
classes to always add a modified version of the
`DefaultArtifactMetadataSource` which attempts to fetch an explicitly
declared artifact if defined. This MetadataSource is added before any
other metadata source so that it will override any other declared
sources.

This was not caught in the past, as the test we used to verify this
functionality was improperly written. It never published the test
artifact without a POM and continued to verify that the POM was being
fetched.

Additionally, clean up some code found along the way by:
 * Adding `@Nullable` annotations where applicable
 * Updating `ComponentOverrideMetadata` to return only a single artifact
   instead of a List since it was only ever initialized with a single
value.
 * Remove `DefaultDependencyArtifact`'s no-arg constructor in favor of
   using the parameterized constructor.
 * Replace usage of `GUtil.isTrue` with `StringUtils.isNotEmpty` when we
   explicitly know a `String` is being used.
 * Avoid potential NPE in `DependencyStringNotationConverter` by using
   `StringUtils.isEmpty`
 * Remove semicolons in Groovy test code.
 * Add unit test to `ModuleFactoryHelper`