Fix potential failure in execution plan health monitoring when using composite builds.
The failure happened when the health monitoring runs while
the last node of a particular plan in the tree is running but not yet complete plus all nodes in all other plans are blocked waiting for  that node to complete.

Also fix a potential race condition finalizing the execution plans of a composite build, by moving the finalization to a single-threaded context (for now).