Add tests for MacOS default native toolchains.
Notes:
1) The modified tests were previously not working, seems that
Spock doesn't evaluate assertions in an interaction.  My solution
of stashing the toolchain in a variable probably isn'y ideal but
I'm open to suggestions.

2) Multi-arch support breaks some previous assumptions,
especially as dummyOS is an instance of the current OS and the
`TargetPlatformConfiguration`s (existing and added) also depend
on the current OS. I worked around it with `@IgnoreIf`s but there
ought to be a way to test all configs on all OSes.

Signed-off-by: Pete Bentley <prb@google.com>