Instrument indy-based Groovy call sites
Modern Groovy versions use Java 7's invokedynamic instruction to do a
dynamic dispatch of methods/constructors/properties. This commits
enables the configuration cache to intercept such dynamic calls and
infer implicit build inputs used in code compiled with invokedynamic
option enabled.

This is a prerequisite to the migration to Groovy 4 as a runtime for the
Gradle build scripts because Groovy 4 has indy as the only option.