Extract registering `javaToolchains` extension to a separate plugin `jvm-toolchains`
Some projects might need to run `java` from a local JVM installation or the JDK tools
while not requiring anything else from `java-base` or `jvm-ecosystem` plugins.
One such use case is running an arbitrary JVM-based tool using a `JavaExec` task,
specifying its `javaL.auncher` using the toolchains feature.

The plugin ID is `jvm-toolchains` and not `java-toolchains`, because the use-case of
using this plugin explicitly is quite exotic, so the plugin should rather not stand
out along with the other user-facing `java-*` plugins. This is similar to the
`jvm-ecosystem` plugin.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>