Validate @Input(File(s)) ResolvedArtifactResult as unsupported
Previously, no validation was in place.

Build would fail on @Input ResolvedArtifactResult at task input fingerprinting stage.
But it would not fail with a task without outputs because then no fingerprinting was happening.

Build would fail on @InputFile because the value could not be resolved to a file.

Build would not fail with @InputFiles, resolving to an empty file collection and disabling task optimisations.

This commit adds validation of @Input, @InputFile and @InputFiles task properties to report usages of ResolvedArtifactResult as errors.

The validation happens both at plugin build time with :validatePlugins and at runtime.

Signed-off-by: Paul Merlin <paul@gradle.com>