Mention help task from help option
Users using the `--help` option don't understand that plug-in
contributed options are not included in the description produced by
`--help`. The help /task/ should be used for that instead.

This adds a highly evident blurb to the output produced by the `--help`
option to direct the user to the help /task/. This changeset also
changes the documentation for `--help` to direct the reader to the
`help` task.

Issue: #19296
Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>