Reuse VersionParser and preserve it across builds
Ensures there is only one globally scoped version parser (promoted from
build scoped) that is shared with all places that need a version parser.

VersionParser keeps a cache of previously parsed versions, to avoid
creating Version objects unnecessarily. However, we had a few places
creating their own long-living copy of VersionParser, and in other
places we where using short-lived VersionParser instances, which
defeated the point of keeping a pool of versions. Also, the existing
shared instance of VersionParser was build-scoped, and this changes
it to be globally scoped so it stays around between builds using the
same daemon, potentially improving performance of the following builds.

Issue: #18144