Set `org.gradle.plugin.api-version` in buildSrc classpath configurations
This attribute was set on the buildscript classpath but
not on the buildSrc runtime classpath (that gets propagated to the main
build).

Set the attribute in the buildSrc classpath configurations, so
that plugins added to buildSrc as dependencies get resolved according to
the attribute.

The attribute is not set in buildSrc subprojects by design: those
projects are by default not used as plugins, so not supposed to build
plugins. If they do, they should normally apply either
`java-gradle-plugin` or `kotlin-dsl` (which applies the former), and that
will handle the attribute for them.