Moving non-null validation to BuiltInOptionElement
Differently than the other AbstractOptionElement subclasses, which are
built based in task options declared by plugin developers,
BuiltInOptionElement is meant to be used in the Gradle codebase alone.
Passing nulls into any of that class' constructor parameters is an
obvious bug, so a NPE is a suitable failure mode, dispensing with the
need to use OptionValidationException.

Issues: #9166