Let the Tooling API control system properties contributed from the client-side
When a client executes a Gradle build invocation via the Tooling API, all system
properties from the client JVM are passed to the daemon. This is not desired,
especially for IntelliJ IDEA, as it uses a large number of system properties that
could potentially break the build (e.g. java.system.class.loader).

This change introduces LongRunningOperation.withSystemProperties(Map) to let the
TAPI client control what system properties are available for the target build
from the client JVM. The exact behavior is defined in
SystemPropertyPropagationCrossVersionTest.

Co-authored-by: nate contino <ncontino@protonmail.com>
Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>