Warn user if task option clashes w/ built-in one
In #9166, we added support for --rerun as a built-in task option.
If a task defines its own -rerun option, the built-in is disabled.
This change aims to tell the user of a task that has an option
clashing with a built-in one (and that the built-in is now unavailable).

Issue: #9166
Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>