Add property `host` to `JavaDebugOptions`: support debugging via network
In Java 9 and above, if the JDWP arguments only include the port and not
the host name, the debugger server will only listen on the loopback
network interface, and it won't accept connections via network.

Debugging remotely with explicitly specified host address was only
possible by passing the JDWP arguments as a string, which is a UX issue.
Adding a property for specifying the host address to `JavaDebugOptions`
fixes that.