Fix wrong type in ImplementationChanges
Previously, ClassImplementationSnapshot was the only known implementation snapshot.
It has changed with the introduction of the LambdaImplementationSnapshot for serializable lambdas.
Now an additional implementation can be either a class or a lambda.
But we don't need to change the type of current implementation because
it is always the class of the task itself.