Fix cycle detection error involving finalized dependencies of finalizer
This is yet another corner case of the finalizer scheduling. It happens
when the finalizer task, say F, finalizes two tasks, say E and D: E is
the entry point and D is a dependency of the finalizer. The
FinalizerGroup created for the finalizer included F and D (which is
correct) and was considering E and D as a successors to the group, so
no tasks of the group could run before them. However, as the group
was contributing to the hard successor edges between nodes, the D task
became its own successor, causing the problem.

The fix tweaks the way finalizer groups are considered when calculating
task successors. Instead of always returning all finalized tasks as
successors for any node in the group, only a subset of tasks is returned
when checking a finalizer dependency. The finalizer still gets all
finalized tasks as hard successors. The subset is carefully to ensure
ordering guarantees for the finalizer dependencies:
1. All finalized nodes outside of the group are kept
2. The nodes that are reachable from the entry point are kept: thes
   nodes are scheduled outside of the finalizer group so a  deadlock is
   unlikely.
3. The nodes that are other finalizers are kept for the same reason.
4. The nodes that are also dependencies of other finalizers are kept
   if the other finalizer is likely to run before this one.
Basically, we strive to remove the finalized group members that are
only added because of this finalizer, as these would cause false
cycles between themselves.

In the example above this would mean that E and D would be
hard successors for F, but only E would be a hard successor for D.
With this the scheduler only relies on the other ordering relations
between tasks when scheduling the tasks inside the finalizer group.

The fixed logic is used both for the planning phase and the execution
phase. The only way where the old behavior is preserved is when
the tasks are checked for cancelability: if the task from the group is
also a finalized task and is already started then finalizer and its
other dependencies should not be cancelled to have a proper cleanup.