Add a reproducer for another false cycle finalizer issue
There the two finalizer groups became successors to each other, even
though it is possible to come up with the schedule that follows all
ordering guarantees. The group A has entryPoint and depB as finalizing
successors, so depB becomes a hard successor for the depA. The group B
has depA as a finalizing successor, so depA becomes a hard successor
for depB. The group-based plan builder consider this a cycle, but it
was possible to run it in 7.4.

The tricky group handling introduced in the previous commit cannot help
here, as the finalized nodes are not in the finalizing groups of their
finalizers.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>