HelpTaskPlugin: Incompatible with Configuration Cache when `taskpath` is specified
It is incompatible because:
Only scheduled tasks are stored into configuration cache,
the specified task follows `--task` would not be scheduled
and won't be part of our stored work graph in most cases.

To get an instance of that task, the plugin tries to reconfigure
the project, but task `Help` is already registered into task
container once configuration cache is loaded, it cannot be
registered again, which causes failure of the reconfiguration.

Fixes #21056

Signed-off-by: Xin Wang <scaventz@outlook.com>