Unify notion of "successors" in cycle detection logic
Previously, the execution plan builder was using hard successors
(dependencies, mustRunAfter, and finalized-by-the-group) to determine
the execution order. If the cycle is detected there, a special
verifier-simplifier is invoked to get a string representation of the
cycle. The verifier was using only dependencies and direct finalized
nodes as successors, so there were discrepancies.

This commit switches the verifier-simplifier to use the same definition
of successor to solve the problem, and removes all unneeded methods.

As a side effect, this turns a "Misdetected cycle" error with finalizer
groups into a proper CircularReferenceException. We've decided to accept
this behavior as a new norm for a while and gather user feedback from
RC.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>