Preserve "stored" compression level when instrumenting JARs
The instrumentation rewrites buildscript classpath JARs, the rewritten
JARs are then used as runtime classpath. The user code sometimes tries
different tricks to find the location of these JARs and use them. One
of the use cases is to memory map a "stored" (uncompressed) resource.
The rewriting code was compressing all JARs content and was breaking
memory mapping. With this commit, stored files will be left stored in
the rewritten JARs as well.

In general, we suggest the users to use a separate configuration to
access non-instrumented JARs instead, but this fix is relatively simple,
so benefits outweigh the cost.