Use more coarse grained UoW.Identity for Kotlin DSL
By using all the execution inputs as the UoW.Identity, the execution history was growing unbounded.
It changed when making changes to the script compilation classpath.

IOW, we stored execution history with a different key each time.

The on-disk history was growing unbounded.
The in-memory cache of this history was growing too fast for our in-memory cache
auto-scaling strategy.

This commits change the UoW identities used for Kotlin DSL to be coarse grained.

Plugin accessors are now keyed by build root dir.
Project accessors are now keyed by project dir.
Script compilations are now keyed by template + filename

Signed-off-by: Paul Merlin <paul@gradle.com>