[WIP] Running the docs tests with the CC enabled
New behavior is enabled in docs:docsTest with a global build flag. It
changes the executor to the one with configuration cache enabled and
toggles the golden output preprocessor mode. The flag is inevitable,
as the docsTest cannot be cloned easily (lots of non-trivial
configuration comes from the binary guides plugin).

The golden output preprocessor allow to add "?? " prefixed lines to
golden outputs. Without CC these lines are removed before the test.
With CC these lines are stripped of prefix and used for the
verification. The approach has its problems though, when the output
is included into the docs. An option could be to use asciidoctor's
tags to strip out this lines from the docs too. Or maybe a custom
include processor that understands "?? " properly could help. There
is also an output normalizer that strips all configuration cache from
the output.

The necessary level of strictness is still unclear: can we just have
"allow unmatched lines" enabled everywhere? Do we care about the CC output?
Do we need to show CC output in the docs.

There is an ignorelist of tests broken because of something other than
output too.