Refactor diff output formats in test fixture
We only need to support a single expected output. This single String
can be a single or multi-line block of text. Refactor the method to
only allow a single String.

Given that we only need a single expected output block, the AUTO diff
output will always chose LEGACY. Remove the AUTO format.

LEGACY is the same verbose output we had before and it's worse in all
scenarios I tried. Remove the LEGACY format.

LINEWISE is better output than UNIFIED. There might be a heuristic that
would let us choose one output over another, but for now, default to
LINEWISE and provide a method to switch to UNIFIED to make it easier to
debug.