Update smoke test to be compatible with Gradle 8.0
AGP and KGP both used deprecated compile options, and needed to be upgraded.  This caused the expected module and pom files used by
this test to change significantly in several ways:

1) Kotlin version numbers needed bumping.
2) Less variants were present (for example - "metadataApiElements-published" gone from "kotlin-multiplatform-android-library-android-demo-1.0.module")
3) Names of existing variants changed (for example - "metadataApiElements-published" became "metadataApiElements" in "kotlin-multiplatform-library-1.0.module")
4) Variant had less attributes present (for example - "com.android.build.api.attributes.VarianttAttr" is gone)
5) Some new attributes present (for example - "artifactType = org.jetbrains.kotlin.klib")

In addition, other properties in the Android extension were removed ("versionCode", "versionName").

But the biggest change necessary was in the comparison logic within this smoke test itself.  The order of the dependencies present in GMM and POM files
doesn't seem to be fixed.  This causes the existing comparisons to fail unless corrected.  This test now sorts the dependencies and then compares the results
of sorting.  Without this code, this test becomes flaky as the order of these nodes in the expected files needs to be changed to match...and then the next time
you run the test, it needs to be changed BACK.