Add build parameters to run docsTest with configuration cache enabled
As we're getting closer to declaring the configuration cache a stable
feature, we have to make sure that our documentation snippets are
compatible with the configuration caching. This commit adds basic
infrastructure of testing doc snippets and samples with the
configuration cache enabled, similar to existing configCacheIntegTest
integration test tasks.

To run the docs test with CC enabled one should specify the property:
```
./gradlew -PenableConfigurationCacheForDocsTests=true docs:docsTest
```
The `subprojects/docs/build.gradle` contains the ignorelist of tests
that are known to be broken. One could run tests from this list only by
adding `-PrunBrokenConfigurationCacheDocsTests=true` to the command
above.

Making a separate task for the docs test with CC was considered, but
turned out to be a very non-trivial task due to the amount of
configuration applied to the `docsTest` by the supporting binary
plugins that has to be replicated.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>