Always apply jvm-ecosystem plugin
The old way was testing from the persepctive of a user which did not apply jvm-ecosyste
which is not supported. In that scenario, the user needs to use internal APIs to apply
the proper schema compatibility rule. No user should be doing this, so we don't need
to test this case.

Therefore, we can delete the half of each test which resolves the variant artifacts
before the java-base plugin is applied. Then, we remove the explicit java-base plugin
application in favor of applying jvm-ecosystem by default, since java-base was in reality
used to apply jvm-ecosystem, the root cause of the original behavior change.