Improve creation of additional variants in JvmEcosystemUtilities
These changes were originally made to support a CLASSES_AND_RESOURCES variant for an
upcoming `compileElements` PR, however plans changed and these changes are not required
for the upcoming PR. Nonetheless, they improve the current state of JvmEcosystemUtilities.

This PR makes the following notable changes:
- Remove JvmPluginExtension/DefaultJvmPluginExtension. This extension is internal and is not
in use in production. It acts simply as a proxy to `JvmPluginServices`. The only use of this
extension was in tests, all of which have been refactored to reference `JvmPluginServices` directly.
In reality, these tests should be referencing public APIs, but updating these tests are out of
scope for this PR. The existence of this plugin hinders further modification/development of
JvmPluginServices, and since it is both internal and not in use, I opted for outright removal.

- Improve `DefaultSourceSetOutput`'s tracking of tasks which contribute to output directories.
The previous implementation tracked all tasks which contributed to _any_ registered classes
directory, requiring a contributing task to be registered in a separate method call than when
the class directory itself is added. This meant we were not able to map which tasks contributed
to which classes directory. This change combines adding the classes dir and the contributing task
into a single method call, thus allowing us to remember which task contributes to which directories.

- Update `DefaultSourceSetOutput` to track which task contributes to the resources output directory.
The previous implementation required out-of-band knowledge of how the resources directory was
generated, however this implementation updates resources to behave similarly to how classes worked
previously, in that the source set output is now aware of the task which contributes to the resources output.

- Update `configureClassesDirectoryVariant` in `JvmEcosystemUtilities` to accept a `Configuration` instead of
the name of the configuration. This way we can avoid a `configurations.all` call which iterated through all
configurations, as each usage of the existing method was called after the configuration was created and thus
we already had access to the configuration in question when making the initial call.

- Add a `configureResourcesDirectoryVariant` to `JvmEcosystemUtilities` to be consistent with the corresponding
method which configures the classes variant. Update `runtimeElements` to use this new method.

- Update documentation of `JvmEcosystemUtilities` to assert this interface is intended for internal use as opposed
to use by plugin authors.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>