Make snippets/dependencyManagement/customizingResolution-attributeSubstitutionRule...
...compatible with the configuration cache.

The task's `doLast` action was reaching out to `configurations` property
of the Gradle script object, causing a reference to this object to be
captured by the action. The configuration cache was attempting to
serialize the action and everything it captures but was failing to do
that because the script object is not serializable.

To work around that a separate `FileCollection` variable `classpath` is
introduced, that is then captured by the `doLast` action. The simpler
type is a hint to the configuration cache to avoid serializing all the
extra stuff the `Configuration` object has compared to the simpler
`FileCollection`.

The changed code is not present in the user documentation and serves
only as a scaffolding, so no changes to the docs are necessary.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>