Split the list of conf-cache incompatible snippets/samples
There are several kinds of failures for which we prefer to have separate
lists, dependending on how soon we plan to address the issues.

Snippets in the configuration cache docs are tested under normal
docsTest run and cannot be tested with the configuration-cache executer
at all. Others can be run with `-PrunBrokenConfigurationCacheDocsTests=true`
if necessary, to check how the failures look like.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>