Move execution deadlock detection up from `DefaultExecutionPlan` to `DefaultPlanExecutor`.
This allows the detection to apply in more cases, for example when all the workers are waiting for work but the execution plan thinks some nodes are running, or when a worker exits unexpectedly while running a node.

This also makes it easier to move other types of work into the executor, as the detection is now independent of the type of work being run and doesn't need to be implemented for each type of work.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>