Make docs:docsTest configuration-cache compatible
The only source of the incompatibility was the
`GradleInstallationForTestEnvironmentProvider` class that was capturing
Task and Project objects. This was replaced with capturing only
necessary state.

The docsTest and its dependencies were also removed from the list of
CC-incompatible tasks. I've checked most of the '*Sample*' and
'*Snippet*' tasks to make sure all of these are actually compatible with
the configuration cache.