Remove javaCrash failure condition and detect daemon expiring message
Sometimes there are JVM crashing and OOM messages in the test log,
which fail the TeamCity build. In these cases, we don't want the build
to fail because they will be handled by test-retry.

However, we do want to capture the OOM in build daemon, let's try
if we can monitor "Expiring Daemon because JVM heap space is exhausted"
message in build log.