Revert "Make Gradle references equal for one referrer project"
This reverts commit 1fa43610ef7421b2276759a78b32c57bef1b8b54.

Reason: after a discussion in the team, we came to a conclusion
that keeping the equality semantics in Gradle references obtained
from different project is a non-goal. It is acceptable to make a
breaking change under project isolation here, given that the use
cases of such equality checks are very likely to violated project
isolation restrictions. For the valid use cases, a new API that
covers them would be more preferable.

Signed-off-by: Adam <897017+aSemy@users.noreply.github.com>